move to dedicated test class
